pyproject.toml
[project]
name = "blogtuner"
version = "0.2.0"
description = "Dead simple static blog generator using markdown files"
readme = "README.md"
authors = [{ name = "All Tuner Labs", email = "hey@alltuner.com" }]
requires-python = ">=3.9"
dependencies = [
    "feedgen>=1.0.0",
    "gitpython>=3.1.44",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "mdformat>=0.7.22",
    "mistune>=3.1.3",
    "pydantic>=2.10.6",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "python-frontmatter>=1.1.0",
    "python-slugify[unidecode]>=8.0.4",
    "toml>=0.10.2",
    "typer>=0.15.2",
]
license = "MIT"
license-files = ["LICENSE"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: Pydantic",
    "Framework :: Pydantic :: 2",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Text Processing :: Markup :: HTML",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Utilities",
    "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/alltuner/blogtuner"
Issues = "https://github.com/alltuner/blogtuner/issues"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true


[tool.bumpversion]
current_version = "0.2.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
ignore_missing_files = false
tag = true
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = true
commit = true
message = "Bump version: {current_version} → {new_version}"
moveable_tags = []
commit_args = ""
setup_hooks = []
pre_commit_hooks = []
post_commit_hooks = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "ruff>=0.11.0",
    "types-python-dateutil>=2.9.0.20241206",
    "types-toml>=0.10.8.20240310",
]

[project.scripts]
blogtuner = "blogtuner.cli:app"

[tool.ruff.lint]
extend-select = [
    # E and F are enabled by default
    'B',      # flake8-bugbear
    'C4',     # flake8-comprehensions
    'C90',    # mccabe
    'I',      # isort
    'N',      # pep8-naming
    'Q',      # flake8-quotes
    'RUF100', # ruff (unused noqa)
    'S',      # flake8-bandit
    'W',      # pycodestyle
]

[tool.ruff.lint.isort]
combine-as-imports = true
lines-after-imports = 2

[tool.mypy]
plugins = ['pydantic.mypy']
